manager sidebar add feedback v1
